--- kobo_book_downloader/globals.py
@@ -1,4 +1,0 @@
-from kobo_book_downloader.settings import Settings
-
-class Globals:
-    Settings = Settings()
