+++ kobodl/globals.py
@@ -1,0 +1,4 @@
+from kobodl.settings import Settings
+
+class Globals:
+    Settings = Settings()
