+++ setup.py
@@ -3,16 +3,18 @@ import os
 from setuptools import find_packages, setup
 
 setup(
-    name='kobo-book-downloader',
+    name='kobodl',
     author='Brandon Davis',
-    version='0.0.1',
-    author_email='kobo-book-downloader@subdavis.com',
+    version='0.1.0',
+    author_email='kobodl@subdavis.com',
+    url="https://github.com/subdavis/kobo-book-downloader",
     packages=find_packages(),
     include_package_data=True,
     install_requires=[
         'click',
         'colorama',
         'dataclasses',
+        'dataclasses-json',
         'flask',
         'pycryptodome',
         'requests',
@@ -20,7 +22,8 @@ setup(
     ],
     license='MIT',
     entry_points={
-        'console_scripts': ['kobodl = kobo_book_downloader:cli'],
+        'console_scripts': ['kobodl = kobodl:cli'],
     },
+    python_requires='>=3.6',
     setup_requires=['setuptools-git'],
 )
